+++ p2n/api.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
 # (c) 2017 The Patent2Net Developers
+import logging
 import p2n.maps
 import p2n.ops.client
 from p2n.model import Patent2NetBrevet
 from p2n.ops.client import OPSClient
-from p2n.ops.model import OPSBiblioSearchResponse, OPSFamilyResponse
+from p2n.ops.model import OPSBiblioSearchResponse, OPSFamilyResponse, OPSRegisterResponse
+
+logger = logging.getLogger(__name__)
 
 
 class Patent2Net:
@@ -18,33 +21,67 @@ class Patent2Net:
         self.documents = []
         self.brevets = []
 
-    def gather(self, expression, with_family=False):
+    def gather(self, expression, with_family=False, with_register=False):
+        """
+        Submit search expression to OPS published search and
+        read response to acquire bibliographic data for each hit.
 
+        Optionally, this will acquire family information for each result document and expand
+        the list of documents by all respective family member documents.
+
+        Also optionally, it will acquire register information for each result document.
+
+        Finally, it will convert the list of result documents in ``self.documents``
+        into a list of dictionaries in legacy Patent2Net Brevet format in ``self.brevets``.
+        """
+
+        # Submit search expression
         data = self.ops_client.search(expression)
+
+        # Debugging
         #print(json.dumps(data))
+
+        # Decode response
         response = OPSBiblioSearchResponse(data)
 
+        # A list of ``OPSExchangeDocument`` object instances
         self.documents = response.results
 
+        # Optionally, expand document list with each documents' family members
         if with_family:
             self.expand_family()
 
+        # Optionally, enrich each document with register information
+        if with_register:
+            self.enrich_register()
+
+        # Finally, provide a list of dictionaries in legacy Patent2Net Brevet format
         self.documents_to_brevets()
 
         return self
 
     def expand_family(self):
+        """
+        Acquire family information for each document in ``self.documents`` and expand the
+        very same list by ``OPSExchangeDocument`` instances of the respective family members.
+        """
+
+        # We are doing bookkeeping using the document's publication number.
+        # Uniqueness checks are performed on this list of publication numbers.
         document_numbers = []
+
+        # The full list of documents including their family members
         documents_expanded = []
+
+        # Iterate all result documents
         for ops_exchange_document in self.documents:
 
+            # Use publication number as key for uniqueness constraint
             document_number = ops_exchange_document.publication_number
 
+            #
             document_numbers.append(document_number)
             documents_expanded.append(ops_exchange_document)
-
-            #print 'document_number:', document_number
-            #continue
 
             data = self.ops_client.family(document_number)
             response = OPSFamilyResponse(data)
@@ -55,9 +92,6 @@ class Patent2Net:
                 if family_member.publication_number == document_number:
                     continue
 
-                #print 'family_member:', family_member.to_json(pretty=True)
-                #print 'yeah'
-
                 # Skip duplicates. Currently uses first-come, first-serve policy.
                 if family_member.publication_number in document_numbers:
                     continue
@@ -67,6 +101,39 @@ class Patent2Net:
                 documents_expanded.append(family_member)
 
         self.documents = documents_expanded
+
+    def enrich_register(self):
+        """
+        Acquire register information for each ``OPSExchangeDocument``
+        object instance in ``self.documents``.
+
+        It will enrich the ``OPSExchangeDocument`` instance by adding more attributes:
+
+        - ``register``: Reference to an instance of ``OPSRegisterDocument``
+        - ``designated_states``: List of country codes of designated states
+        """
+
+        # Iterate all result documents
+        for document in self.documents:
+            document_number = document.publication_number
+            try:
+
+                # Fetch register information from OPS
+                data = self.ops_client.register(document_number)
+
+                # Read register information response
+                response = OPSRegisterResponse(data)
+                if response.results:
+
+                    # Is there more than one register document sometimes?
+                    register_document = response.results[0]
+
+                    # Propagate register information into OPSExchangeDocument object
+                    document.register = register_document
+                    document.designated_states = register_document.designated_states
+
+            except:
+                logger.debug('No register information for document "{}"'.format(document_number))
 
     def documents_to_brevets(self):
         """