+++ p2n/ops/model.py
@@ -2,12 +2,19 @@
 # (c) 2017 The Patent2Net Developers
 import re
 import json
+import logging
 from collections import OrderedDict
 from jsonpointer import JsonPointer, JsonPointerException
 from p2n.util import to_list
 
+logger = logging.getLogger(__name__)
+
 
 class OPSBiblioSearchResponse:
+    """
+    Read the response from OPS published data search and decode the
+    search results to a list of OPSExchangeDocument objects.
+    """
 
     def __init__(self, data):
         self.data = data
@@ -24,6 +31,10 @@ class OPSBiblioSearchResponse:
 
 
 class OPSFamilyResponse:
+    """
+    Read the response from OPS family retrieval and decode the
+    results to a list of OPSExchangeDocument objects.
+    """
 
     def __init__(self, data):
         self.data = data
@@ -34,28 +45,72 @@ class OPSFamilyResponse:
         pointer_results = JsonPointer('/ops:world-patent-data/ops:patent-family/ops:family-member')
         family_members = to_list(pointer_results.resolve(self.data))
         for family_member in family_members:
+
+            # Decode document number
+            publication_number = 'unknown'
+            try:
+                document_id = JsonPointer('/publication-reference/document-id')
+                publication_number, publication_date = OPSExchangeDocument.decode_document_number_date(document_id.resolve(family_member), 'epodoc')
+            except:
+                pass
+
+            # Read bibliographic data for family member
             item = OPSExchangeDocument()
             try:
                 item.read(family_member)
                 self.results.append(item)
             except JsonPointerException:
-                # FIXME
-                pass
+                logger.warning('No bibliographic data for family member "{}"'.format(publication_number))
+
+
+class OPSRegisterResponse:
+    """
+    Read the response from OPS register information retrieval and decode the
+    results to a list of OPSRegisterDocument objects.
+    """
+
+    def __init__(self, data):
+        self.data = data
+        self.results = []
+        self.read()
+
+    def read(self):
+        pointer_results = JsonPointer('/ops:world-patent-data/ops:register-search/reg:register-documents')
+        register_documents = to_list(pointer_results.resolve(self.data))
+        for register_document in register_documents:
+            item = OPSRegisterDocument()
+            try:
+                item.read(register_document)
+                self.results.append(item)
+            except JsonPointerException:
+                logger.warning('Could not read register information from data "{}"'.format(register_document))
 
 
 class OPSExchangeDocument:
+    """
+    Implement the data model of OPS exchange documents,
+    optionally enriched by register information.
+    """
 
     def __init__(self):
+
+        # Bibliographic data
         self.application_date = None
         self.application_number = None
         self.publication_date = None
         self.publication_number = None
+        self.country = None
         self.title = {}
         self.abstract = None
         self.applicants = []
         self.inventors = []
 
+        # Register information
+        self.register = None
+        self.designated_states = []
+
     def to_json(self, pretty=False):
+        """Convert document to JSON format"""
         if pretty:
             return json.dumps(self.__dict__, indent=4)
         else:
@@ -63,6 +118,10 @@ class OPSExchangeDocument:
 
     @staticmethod
     def decode_document_number_date(docref, id_type):
+        """
+        Decode document number and filing/grant date from
+        /exchange-document/bibliographic-data/{application|publication}-reference/document-id.
+        """
         docref_list = to_list(docref)
         for document_id in docref_list:
             if document_id['@document-id-type'] == id_type:
@@ -76,6 +135,9 @@ class OPSExchangeDocument:
 
     @staticmethod
     def decode_titles(titles):
+        """
+        Decode titles in all languages.
+        """
         data = OrderedDict()
         for title in titles:
             language = title.get(u'@lang', u'ol')
@@ -86,6 +148,9 @@ class OPSExchangeDocument:
 
     @staticmethod
     def decode_abstracts(abstracts):
+        """
+        Decode abstracts in all languages.
+        """
         data = OrderedDict()
         for abstract in abstracts:
             language = abstract.get(u'@lang', u'ol')
@@ -101,6 +166,10 @@ class OPSExchangeDocument:
 
     @staticmethod
     def decode_parties(partylist, name):
+        """
+        Decode list of applicants or inventors.
+        """
+        #print 'partylist:', partylist
         parties = []
         for party in partylist:
 
@@ -115,14 +184,17 @@ class OPSExchangeDocument:
         return parties
 
     def read(self, data):
+        """
+        Read information from data object.
+        """
 
+        pointer_country = JsonPointer('/exchange-document/@country')
         pointer_application_reference = JsonPointer('/exchange-document/bibliographic-data/application-reference/document-id')
         pointer_publication_reference = JsonPointer('/exchange-document/bibliographic-data/publication-reference/document-id')
         pointer_invention_title = JsonPointer('/exchange-document/bibliographic-data/invention-title')
         pointer_abstract = JsonPointer('/exchange-document/abstract')
         pointer_applicant = JsonPointer('/exchange-document/bibliographic-data/parties/applicants/applicant')
         pointer_inventor = JsonPointer('/exchange-document/bibliographic-data/parties/inventors/inventor')
-
 
         pubref = pointer_publication_reference.resolve(data)
         pubref_number, pubref_date = self.decode_document_number_date(pubref, 'epodoc')
@@ -156,6 +228,7 @@ class OPSExchangeDocument:
         except JsonPointerException:
             inventors = []
 
+        self.country = pointer_country.resolve(data)
         self.application_number = appref_number
         self.application_date = appref_date
         self.publication_number = pubref_number
@@ -164,3 +237,33 @@ class OPSExchangeDocument:
         self.title = title
         self.applicants = applicants
         self.inventors = inventors
+
+
+class OPSRegisterDocument:
+    """
+    Implement the data model of OPS register documents.
+    """
+
+    def __init__(self):
+        self.designated_states = []
+
+    def read(self, data):
+
+        designated_states_reference = JsonPointer('/reg:register-document/reg:bibliographic-data/reg:designation-of-states')
+        self.designated_states = self.decode_designated_states(designated_states_reference.resolve(data))
+
+    @staticmethod
+    def decode_designated_states(data):
+        """
+        Decode designated states from register document.
+        """
+
+        countries_reference = JsonPointer('/reg:designation-pct/reg:regional/reg:country')
+        countries_raw = to_list(countries_reference.resolve(data))
+
+        states = []
+        for country_raw in countries_raw:
+            country = country_raw['$']
+            states.append(country)
+
+        return states